refactor: create store dir
